backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.1.1
PyJWT==2.11.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
Django==5.2.0
django-cors-headers==4.4.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
psycopg2-binary==2.9.11
python-dotenv==1.0.1
requests==2.32.3
drf-spectacular==0.29.0
gunicorn==22.0.0
whitenoise==6.7.0
dj-database-url==2.2.0
reportlab==4.2.2
uritemplate==4.2.0
urllib3==2.6.3
